+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Added the `Application.exception_handler` decorator which registers exception handler functions.
+
 ## [1.0.3] - 2021-04-24 :cookie:
 - Adds support for Python 3.10 and PEP 563 (however, it works with `httptools`
   built from its current default branch, because the version of `httptools`